server/requirements.txt
Flask
flask-cors
redis
python-dotenv